RadioButton: make sure focus is not lost when pressing a key stroke
If a radio button is focused and a key stroke is pressed to select
another radio button in the same group, the newly selected button should
gain the focus because the deselected button will lose it due to the
removal of its tab index. If this were not done the body would gain the
focus.

Also added getFocusableElement to Widget.js so that focus() and
isFocused() can work with the same element.