isContentTruncated should also work if current element has no size.
Before the isContentTruncated method required that the size of an
element is defined by the element itself to work. In case the size is
given by a parent element, always false was returned (in
Internet-Explorer).
Now the algorithm ensures that the parent has no effect and allows the
child to grow as required by its content. This allows to compare the
actual width of the element with the width as it would be if not
restricted and therefore only defined by its content.

#217831