SmartField: always set display text (also in case of a VetoException)
When the user types something, the smart-field should keep that text,
even when a VetoException is thrown in execValidateValue.

Also fixed a small bug which occurred when the user types an invalid
text, presses TAB and then clicks on the clear icon. Before the bugfix
the next field was focused after the click, which is wrong.

222979