UiServlet: options request must never return 404
Otherwise browser throws an error saying 'Response for preflight has
invalid HTTP status code 404'.

Also for head and trace the HttpServlet seems to have a decent default
implementation, so it is fine to call it if the request is not handled
by any handler.

Co author: claudio.guglielmo@bsiag.com

Change-Id: Ib7e98a9cfb04b5f82bb49abb00297eda3208a2a5
Reviewed-on: https://git.eclipse.org/r/119503
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>