Accordion: support form fields as group body
- The FormFieldLayout positions its content absolutely. If the form
field is used outside of a logical grid (e.g. in the group body) the
content is mispositioned. Fixed by adding position: relative to the
.form-field class
- If the group body is an image, the image will be loaded during the
opening animation. This means the preferred size is wrong initially and
the body should be resized to the new size as soon as possible.
- The change in html component allows a widget to react to layout
invalidations triggered by a child.