FormFieldTile: add display style
The form field tiles are designed to be used on a dashboard. If you just
want to use them as container for a form field, the dashboard style may
be annoying. To allow that use case a display style 'plain' has been
added. A form field tile in that mode basically just renders the form
field without adjusting the look or behavior.