ClientNotificationPoller adds cancellable which is never unregistered
Previously in commit 55768b274d8e2a94971af5a23cde882e4fc7e22f a new
RunMonitorCancellableProcessor was introduced to just register the run
monitor to its parent during the actual run (and unregister it
immediately afterwards).

This changed the behaviour of the JobFutureTask as it was not cancelled
immediately anymore the parent run monitor was cancelled (if it was
cancelled before the actual run), now its run monitor is cancelled
during the callable chain of the future. In the past the future was not
run if the parent run monitor was cancelled prematurely.

To maintain this behaviour with the new RunMonitorCancellableProcessor
an additional callable chain interceptor (and corresponding test) is
introduced which stops the callable chain if the run monitor is
cancelled.

222822

Change-Id: I636de8978a331f5182d61fc91d41d5641aa2376c
Reviewed-on: https://git.eclipse.org/r/119701
Tested-by: CI Bot
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>
Reviewed-by: Matthias Otterbach <matthias.otterbach@bsi-software.com>