LoadingSupport: skip delay if widgets is initially loading
If widget is loading while being rendered, the loading indicator delay
should be ignored. Otherwise, the empty widget may be visible for a
short amount of time, only to be hidden again after the delay. The
loading indicator delay should only be applied when the widget is
already rendered when it starts loading.

Change-Id: If755ad2acecb3ed08ef376bf41a7c9ff8f1748b7
Reviewed-on: https://git.eclipse.org/r/120096
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>