FormField: improve preferred size calculation
FormFieldLayout now respects the widthHint and heightHint and adjusts
them as needed. This is especially necessary for form fields having
content with variable height which depend on the width of the parent
(e.g. LabelField).

Also added a min-height rule for input fields. This ensures that form
fields with an input element are always at least as big as a logical
grid row. This is necessary to have a consistent look and to make sure
status icon is aligned properly with the field. For other form fields
this may not be true, so now it is possible to have a html or label
field smaller than logical grid row height, at least if the label is
invisible. If this is not wished just add a min-height rule for the
specific field.

Cleaned up implementation and added specs.