FormField: allow dynamic change of statusPosition and labelPosition
When the label position changed from top to left it did not work because
the logicalGridData still had the logicalRowHeightAddition set.

Also adjusting statusPosition on the fly did not work because $status
still had line-height set.

Furthermore if an error status or tooltip was toggled when status
position was top the icon was positioned wrongly because toggling the
tooltip did not trigger a relayout. It still does not trigger one but
instead the status is now always 16px height even if there is no icon
shown yet.