Scout JS: Add automatic layout for SequenceBox & RadioButtonGroup
In Scout JS radio buttons in a RadioButtonGroup and form fields in
SequenceBoxes are now layouted automatically. There is no need to
manually set the grid positions.
Furthermore the RadioButtonGroup has a new property 'gridColumnCount' to
control how many columns that should be used to layout radio buttons. If
not specified the columns are created as need to show all buttons within
the height of the group. Otherwise the specified number of columns is
used.