LogicalGridLayout: improve useUiHeight support by supporting widthHint
Some components may have content with variable height, e.g. label fields
with wrap text = true. In order to calculate that preferred height
correctly, they need a width hint. This width hint has to be passed from
a component width a fixed width.

Because such form fields are mostly used in a group box with a logical
grid layout, the logical grid layout has to pass the width hint if the
form field wants to be as big as its content (useUiSize = true). In
order to do this the logical grid layout needs to know how wide the
individual fields will be. Because this depends on the configured grid
datas (e.g. weightX, widthHint, gridw), the widths have to be calculated
before the preferred heights can be read.

Also improved options handling of HtmlComponent.prefSize function to be
consistent with scout.graphics.prefSize. Until now it was not clear
which options are possible, especially includeMargin option.