Desktop: render browser history entry initially (deep link)
When the desktop is rendered initially, the property
PROP_BROWSER_HISTORY_ENTRY should be rendered. Otherwise, when reloading
the page (Ctrl-R), the model might unexpectedly reset the current view
to the default startup view, because no deep link is present. When a
startup bookmark is activated, this should be reflected in the UI.

When the history state is rendered initially, it should "replace" the
current state instead of adding an entry to it. Otherwise, the use could
click "back". The initial state does not always contain a deep link, and
when it is missing, nothing would happen.

Change-Id: Icbb4fdedd51616c04098091ee8c826301b333615
Reviewed-on: https://git.eclipse.org/r/120097
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>