Enabled state must be inherited even if connected to the Java model.
The server-side Java model only sends enabled-property update events for
the root element that changed. This minimizes the data transferred.
Hence the enabledComputed state must be recalculated and inherited on
the JavaScript side to update all child widgets.
Furthermore on the Java model enabled state is only supported for
FormFields. But the inheritance must also be propagated to non-FormField
widgets.