Include button-disabled special cases in ButtonAdapterMenu
There is a special logic that system-buttons of type close or cancel are
not disabled if the container is set to enabled-granted=false. This
allows to close the form even if all fields are disabled
(read-only-mode). The same logic applies for the ButtonAdapterMenu which
should behave exactly as the wrapped button. Therefore delegate the
enabled-computed calculation to the wrapped button instead of using the
default logic for widgets.