Widget: always remove children before the parent is removed
If the child is removed using an animation it may happen that the parent
is removed while the animation is still running. In that case the child
is removed after the parent, which is different to the regular remove
order. This can have strange effects, e.g. scrollbars throwing an
exception because they try to read a data of a jquery object
($parent.data('scrollbars')) which does not work if the parent has been
removed already.