ContextMenu: last menu item is too big sometimes
When the individual menu items have a fractional height, the preferred
size of the popup will be too big because the values will be rounded up.
When the popup is layouted the items won't get an absolute height
because ContextMenuPopup disables pixelBasedSizing of the RowLayout.

To solve this the preferred size needs to return exact values.