ContextMenu: items are not fully visible anymore if sub menu is closed
If a sub menu is opened and closed again, the context menu has the wrong
height. Reason: the sub menu title item ($placeholder) is not linked to
a html component, so the RowLayout cannot calculate the preferred height
correctly. This worked before the LogicalGridLayout change by accident,
because the preferred size of the sub menu was cached. This size is not
cached anymore because now a widthHint may be set by the
RowLayout.layout which is considered when caching.