Fix layout of fields in menu bar / context menu
- StringField was not visible anymore in the menu bar
Reason: useCssSize set in MenubarBoxLayout, ColumnLayout etc. is now
propagated to the actual field. But since the field is configured to use
100% of the container width, it returns 0 as preferred width. This means
percentage values cannot be used when preferred size is calculated. To
solve it, the 100% rule is only added for ie 9, since it is a hack for
ie 9 anyways.

- DateField had the wrong in the menubar
Reason: DateTimeComposite returned the wrong pref height, min-height is
configured at input level.

- DateField was too big if opened as single menu in the context menu
Reason: min-width had no effect because it was set on the composite.
This is now supported by the DateTimeCompositeLayout

- Small improvements of preferred size calculation of form field layout
If useUiHeight is set and field itself returns a smaller ui height than
label and status, the height of the label and status should be used as
preferred height. Otherwise the container will be smaller than label and
status. This may be the case for date fields if hasTime and hasDate is
set to false.