Refactored EventListenerList
The class EventListenerList has poor performance with large numbers of
listeners and add / remove operations.
There are two new alternatives to this class that are also thread-safe
and support higher performance.
* FastListenerList<LISTENER> is used to manage a single type of
listeners
* AbstractGroupedListenerList is used as base class to handle a
single type of listener with multiple type partitions. See the new
classes TreeListeners and TableListeners for an example of
applicability.

Consider refactoring the use of EventListenerList by one of the new
alternatives.

Change-Id: Ied82df5399b8fa9a1440e42db7757b4bd223c3cc
Signed-off-by: Ivan Motsch <ivan.motsch@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/121012
Tested-by: CI Bot