TileGrid: don't throw exception if gridData is null
There may be widgets without a gridData. Since the gridData is required
by the layout it will throw an exception. This can happen if the widgets
are inserted and removed using an animation before the layout has been
done. If the widget is removed using an animation it is not in the list
of getGridWidgets() anymore but may still be in the DOM.

Stacktrace:
Uncaught TypeError: Cannot read property 'x' of null at
TypeError: Cannot read property 'x' of null
    at scout.LogicalGridData.validate
    at validateGridData

This also fixes a problem that the group in a tile accordion hat the
wrong height after insertions and deletions of groups and tiles. The
layout layouted removed tiles which were still in the DOM due to the
animation.