Delete global disabled color style
Instead having a global default disabled color style only explicit
disabled styling should be used. This has the advantage that new
components do not get disabled styling automatically and that the style
must not be removed in case it is not desired. Therefore the global
style was deleted and explicitly used where it was active before.