Fix flickering of labels when shown on a beamer
If the app is shown on a beamer with Chrome, every label on the screen
flickers when a popup is opened (e.g. smart field popup). This did not
happen with scout 7.0, the reason seems to be the new position: fixed
rule for the body tag which was introduced to prevent body scrolling on
ios.

Fixed by only applying the rule on ios devices.