fix enabled bug
in Form.js the enabled Method of Widget.js was overwritten with another
set of input parameters. Therefore the whole form was set as enabled
when one tries to set only one Field as enabled. The convenience method
"enabled" in Form.js isn't needed any more, since it is now available in
the super class Widget.js.

Change-Id: I5de51f5a22efb7389274ddb78d233b56c99cc3e5
Signed-off-by: Nicola Steger <nicola.steger@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/120991
Tested-by: CI Bot
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>