Improve Action.js
Action.js was mainly used as abstract class for Menu and other widgets
and did not implement a render function. This has been changed so that
Actions can be used as a regular widget. The CSS rules are kept simple
so far, the widget is not fully featured yet. It may be enhanced in the
future.

Also added 2 new features:
- The showing of the tooltip will be canceled when the action is
clicked. It was annoying if the user wants to click the action but
instead the tooltip is shown. If he wants to see the tooltip he has to
hover over the action and wait a little.
- A new property has been added called showTooltipWhenSelected. This
behavior is required for example if the tooltip changes when the
selection changes. The behavior may not be desired if the action toggles
a popup (e.g. TableControl, Menu with popup, view button)