getForm() of child must always return the same as getForm() of parent
Since 8.0, the form is set in initConfig by using a thread local
(IForm.CURRENT). That means when creating a dynamic form field, the form
may be already set even though the field is not attached to a form yet.
When the field is attached to a form, getForm() needs to return that
form and not the previous one.

Concrete case: JsonButton uses getForm() and returned the outer form
instead of the inner form. This resulted in an exception because
ButtonAdapter.js also uses getForm() but got the inner form. Hence, the
key stroke scope could not be resolved -> exception.