Button: improve keyStrokeScope
1. No need to send the form as scope to the browser because Button.js
can find the form by itself
2. Button.js now resolves the id instead of the ButtonAdapter.js. This
has the advantage that the scope can be defined in a model based json
file when using Scout JS.
3. The outermost form is now used to resolve the id instead of the
innermost. The use case to enable the key stroke for a dialog even if
the button with the key stroke is contained in an inner form seems
reasonable to me.