Fix junit test
Some tests expected that the form is still set after a field is removed
from a composite. The main reason is to prevent the moving of a field
from one form to another. This is currently not allowed because most
form fields are inner classes. Such form fields still point to the
original outer class after the movement. If the code (e.g. in an exec
method of the form field) relies on that it may break.

-> revert to the former behavior