Desktop: always bring outline to front when activated
When a different outline is activated, the desktop will usually bring it
to front automatically. However, when a form is currently activate and
the outline to activate is already selected, the outline would not come
to front. To make sure the behavior is the same, the corresponding
"outline content activate" event is fired manually.

224881

Change-Id: I90ca76e0dc55b1bc27d626e9f103853d0e2d3023
Signed-off-by: Nicola Steger <nicola.steger@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/121327
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>