TabItem: show tooltip if there is an ellipsis
The TabItem does not show an ellipsis by default, but the developer may
add a css rule (e.g. max-width on the .title element). If he does he
would have to adjust the JavaScript code as well. To make that easier,
TabBox.js now supports that case so that adjusting CSS is sufficient.