HtmlComponent: consider min/max width/height
If a component sets one of these constraints using CSS the layout should
consider it when calculating the preferred size. Otherwise the component
itself would be reduced to that size correctly but the dialog may be too
big.

Restriction: these constraints can only be pixel based. A percentage value depends on the container size which is not available at that moment. If a percentage value is used unexpected sizes will be returned -> don't use it.