Only execute LookupCalls once per instance
A LookupCall is only executed once. For each new call a fresh clone of
the template LookupCall is created. This is necessary so that a
LookupCall can have call specific state.
Furthermore a new event 'prepareLookupCall' has been added to SmartField
and TagField that allows to edit the LookupCall before it is executed.
This allows to set custom properties/arguments to a LookupCall that may
be necessary for the call.