Form: improve open() and add when functions for promise chains
- form.open() now loads the data first before showing the form. This has
the advantage that the user does not see an empty form first before the
data is loaded. This means the form is now always rendered later and it
is not possible to access ui properties (e.g. $container) right after
calling open(). Instead form.open().then(...) has to be used.

- added form.whenClose, form.whenSave etc. for developers who prefer
promises over callback functions. Instead of form.on('close',
function(){}) it is now possible to use
form.whenClose().then(function(){})

- also added the generic function when(type) to Widget.js to allow that
pattern for every widget.