TileAccordion: add filter functions to make filtering of tiles easier
Added: addTileFilter, removeTileFilter, setTileFilters, filterTiles

Also fixed filter animation when the group is made invisible while
filtering (the animate-invisible css class is not removed if the group
is made invisible while filtering. When the group was made visible again
the invisible animation was executed instead of the visible one).

Also: tileAccordion now triggers property change events for 'tiles' and 'filterdTiles' as well.