Table: size of empty data is not updated on column structure change
If a column is removed or added when there is no data, the scroll width
of the table data was wrong. Because of that the organize menu was
positioned wrongly and may even be moved out of sight when the table was
horizontally scrolled to the left after a column had been removed while
the table was scrolled to the very right.

Also, empty data was not rendered when all rows were deleted.

214273