Bug 534293 Fix ServerRunContextProducer.produce(Subject)
Previously this method compared serverRunContext.getSubject() with the
subject it just set there, this comparison always evaluated to true
(resp. false because the comparison was negated). As we cannot access
the subject of a session directly (session does not store it) we compare
the current subject with the new subject (as the current subject should
be the subject of the current session, we just copied the current run
context and set the session of the new run context to the current
session).

226681

Change-Id: I7d7c62be220c23e502428b22193083ef2c3ca288
Reviewed-on: https://git.eclipse.org/r/122076
Tested-by: CI Bot
Reviewed-by: Matthias Otterbach <matthias.otterbach@bsi-software.com>