Accordion: improve animation by considering paddings
If the body had a padding, the collapse animation was not smooth because
the group only shrinked to the size of the padding using an animation
and then stopped abruptly.

To fix that, the paddings are reduced to 0 as well. This created the
problem, that the tiles were moved when expanding, but not when
collapsing. Reason: The body was layouted while expanding (by
GroupLayout.layout) and the current padding state was (correctly) taken
into account by the TileGridLayout.

To fix that, the layout was only done once before the animation starts
using the target size. This should also improve the performance.

Important: In some cases the body layouting still needs to be done during the animation, e.g. if an image loads while the group is being expanded). This still works because group.resizeBody updates the size of the body using setSize when GroupLayout.invalidate is called.