Outline: prevent unnecessary animation when calling navigateToTop()
The method outline.navigateToTop() should reset the expansion state of
the tree to the initial state. When first calling "collapse all" and set
only expanding the nodes with initialExpanded=true, this will result in
an unnecessary "expansion" animation. Instead, each node should have
it's initialExpanded value applied, so when the state is already
correct, no animation is visible.

Additional change in TreeNode.js:
Initialize the property "initialExpanded" with the value of "expanded"
when it has no explicit value specified by the model.
This is similar to the Java implementation of AbstractTreeNode.java
(getConfiguredExpanded).

Change-Id: I949965acdef84804178835015e4219a323c063c4