Table: prevent JS error when clicking with both mouse buttons
When both mouse buttons are pressed together on a table header
separator, a JS error occured, because the same event listener was fired
twice, but the referenced DOM elements were already gone after the first
one. Fix: moving the separator or the header item should only be
possible when pressing the left mouse button.

Change-Id: Id132f64d035e2f5c4b800bc238be4de6acf95980