Tile: remove default selection style
The selection style depends on the style of the tile. The current
selection style needs to add a padding to the tile grid so that the
shadow won't be clipped. This may not be desired because the tiles are
not aligned on the left side anymore. Also, when the selection style
needs to be changed, one has to remove several css rules which is a
little annoying. Having a default selection style is not easy because
there is no default style of a tile either. Current solution is to
remove the selection style completely.