Tree: fix wrong node order in "flat list" on node expansion
When a node is inserted into the "flat list" of visible nodes (virtual
scrolling), it automatically adds all of it's children to the flat list
as well. When there are no children, the insertion index of the current
"insert batch" was altered, which then resulted in completely wrong node
orders. This caused wrong displays of tree, as well as possible JS
errors when scrolling, because the "view range" indexes did not match
with the actual node indexes. To fix this problem, the insertBatch index
must not be changed when there are no child nodes.

226275

Change-Id: I9a87cb44cfc6709bc5398f70ca6685e4dca66679