Table: fix resizeColumn() for invisible columns
To find the correct DIV to resize when the table contains invisible
columns (with visible=false or displayable=false), the visibleColumns()
array must be used to calculate column indices.

Relevant use cases:
- A column is invisible and the user manually resizes a column to the
  right of this column.
- A column is invisible and has visible columns to the right and the
  table has "autoResizeColumns" set to true.
(This applies only to Scout JS, because when columns are invisible in
Scout Classic, the hidden columns are deleted from the UI entirely.)

Includes a Jasmine test.

[Minor]
Cell.js: remove obsolete $cell property.

Change-Id: I8e9852bc5fd2b218d6e93de4a11a8f500b4b884f
Reviewed-on: https://git.eclipse.org/r/122787
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>