Tree: fix another node insertion bug
When an insertBatch during node expansion contains nodes of different
levels, the calculated insert position was sometimes incorrect. When an
insertBatch is not empty, the insert position should not be changed.
Otherwise, a previously correct insert position might be overwritten by
a wrong value if the current node is on a different tree level than the
first node in the insertBatch.

227209

Change-Id: I43665cf080e53bd9157ac220c3ec41bc25e901b0