FormField: improve top label style
With the new alternative style, the top label seems to be very far away
from the actual input. It is not clear which label belongs to which
input especially if the input is empty.

The new label is smaller, has no bottom padding and is grey instead of
black.

The form fields which still have a border (list boxes, multiline string
fields etc.) still have a bottom padding, otherwise the label would be
too near to the top border. Small drawback: Not all inputs are aligned
anymore. Example: a multiline text on the left, a single line text on
the right. The labels are aligned, the text inside the inputs aren't
anymore.

Also:
 - Moved _onFieldFocus and _onFieldBlur from ValueField to FormField
because every form field has to set the focused class on the container.
 - Moved _onFieldInput from ValueField to the fields using it. This is
more consistent, otherwise ValueField would define a function which
itself never calls.