RadioButtonGroupGrid must calculate grid based on column only.
The RadioButtonGroup has a property 'gridColumnCount'. The grid of the
group should be calculated based on this property only. The number of
rows corresponds to ceil(buttons/columns).
The height of the RadioButtonGroup must not be considered anymore as
this is done in the default calculation of the column property (for
legacy reasons where the grid was controlled using the field height).

Change-Id: Iba46ab42f54a1ec1ca46a3fe64e2d11031be0761
Reviewed-on: https://git.eclipse.org/r/122940
Tested-by: CI Bot
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>