Align Tree collapse/expand icons
The collapse/expand icon of the tree should be aligned to the text of
the parent node. In case the tree is checkable this is the center of the
checkbox of the parent node. otherwise its left aligned to the text (or
icon) of the parent node.
The collapse/expand icons change color on hover.