Accordion: improve styling and usability
- Expand collapse on mouse down rather than click -> faster interaction
- Removed top border on first group because it is unnecessary and
disturbing in many cases
- Added new collapseStyle property with 2 styles
  1. left: collapse icon is aligned left and looks like a tree node
control (arrow points right and down).
  2. right: collapse icon is aligned right and tile suffix as well.
Arrow points down and up.

This two modes are necessary because on small accordions the right style
looks better than the new default left style. The left style looks
better on big accordions because the arrow and the tile suffix won't be
too far to the right and therefore visible to the user.