DateField: selecting and invalid date possible by clicking twice
If an errorStatus is set in execChangedValue, picking that date leading
to the error using the date picker will set the error status correctly,
but only the first time. If the same date is clicked a second time the
error status will be removed. This does not happen if the validation
code is executed by execValidateValue.

Reason: when a date is picked, field.setValue is called. The
DateFieldAdapter used a parser which was not executed in that case.
Changing it to a validator solves that problem.

Unfortunately, this was not sufficient. When an ui error status was set
in the meantime the server sent the ui error status back. The adapter
stored that status incl. its display text which was not a parsable date.
When the not allowed date was picked, the new display text did not match
the display text of the error status anymore, therefore the model error
status could not be reverted. The server now sends the UI and the model
status back and the adapter ignores the UI status.

226855