Fix listeners: ensure that order of listener is preserved
When adding a listeners to all events (empty vararg) and another to a
specific event type A then the order of adding to event type A is not
preserved.

In addition, when adding a listener to all events (empty vararg) and
also to event type A then the listener was called twice upon fireEvent
A, once for event type A and once for the 'allEventType'.

This fix changes the concept of adding an all-event-type-listener.
Before: add Listener to the allEventsType() listener list
After: add Listener to all knownEventTypes() listener lists and also add
the listener to the otherEventsType() listener list
Change-Id: I318a26908d3765c7c40f043cf273d09fcf58dee9
Reviewed-on: https://git.eclipse.org/r/123614
Tested-by: CI Bot
Reviewed-by: Ivan Motsch <ivan.motsch@bsiag.com>