Tile: add displayStyle with a default style
Moved displayStyle from IFormFieldTile to ITile and added a new option
called DEFAULT. If the default style is active, default CSS rules are
applied (some padding, border, selection style etc.). This makes it
easier to use styles without the need to customize it completely.

Customizing is still possible. Just set the display style to PLAIN and
it will behave as before.