Show connecting message longer
If the connection to the server is lost the offline notification is
shown. While a reconnect attempt is running the message 'Connecting...'
is shown. If the attempt fails very quickly, the message cannot be read
because it disappears too fast. Therefore show the message longer even
it is already clear that the attempt failed.