ButtonAdapterMenu.adaptButtonProperties: Respect defaultButton property
Initially map defaultButton -> defaultMenu only if defaultButton is set
to true, for other cases initially keep  defaultMenu = null as default
value (also if defaultButton is set to false) as the
scout.MenuBar.prototype.updateDefaultMenu method would not update the
defaultMenu according to the keyStroke. For property changes always
update the defaultMenu property according to the defaultButton setting.

228922

Change-Id: Ie9a5294e5ab1e7b6fb23bace3b99bcd76d3b372a
Reviewed-on: https://git.eclipse.org/r/124083
Tested-by: CI Bot
Reviewed-by: Matthias Otterbach <matthias.otterbach@bsi-software.com>