Ignore JS only table user filters
If a TableUserFilter is added to a JS table the model is informed about
the new filter using the event 'filterRemoved' or 'filterAdded'. But not
every filter type has a model-side Java representation. If the filter
type is not supported. Ignore them on the model.
For now only text filters and column filters exist (see java hierarchy
of IUserFilterState).