TabBox: adjust focus marker again
The current focus marker is a triangle on the line pointing to the text.
It is confusing to the user that this triangle vanishes and appears
again suddenly. It is not clear that it reflects the focus state. The
new marker just colorizes the selection marker a bit which is hopefully
less confusing.