Table: context menu key strokes don't work every time
Reason: focus was set to table by the mouse down after the popup had
been opened because event.preventDefault() had not been called because
the mouse down listener did not return the result of the selection
handler. It happened only when the request returned very fast (mostly
when right clicking the same row again). It could be reproduced every
time when the context menu was shown immediately without waiting for the
request to return.

224193