Rename UiThemeUtility to UiThemeHelper and upgrade to a bean
Furthermore a validation method has been added. This allows projects to
replace the utility and apply custom validation logic of the themes.

This may be useful if theme names change and a mapping old->new is
required. Or to delegate to a default theme if the theme name may not be
correct.