Form: position $header absolutely when title is missing
When a form has no title and subTitle, the $header should not use any
vertical space. Otherwise, a large white "gap" would be visible at the
top of the form.

Instead, the $header is positioned absolutely (effectively taking it out
of the normal element flow) by adding a CSS class "no-title". The
FormLayout checks for this property when calculating the header size.

Change-Id: I8e40665241aecf87222e1ce501cdbe12a8bbccab
Reviewed-on: https://git.eclipse.org/r/125088
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>