Improve outline breadcrumb mode
- Removed arrow symbols from ancestors. This style does not work well if
the navigation uses a gradient as background.
- Used same background-color as outline title for ancestors. This should
make it more clear that the outline-title is the root node.
- Made lines darker, made ancestor lines even darker to more clearly
separate the ancestors from the children of the selected node.
- Added active state to outline-title to provide touch feedback. Only
enabled on touch devices because it may be annoying when clicking with
the mouse. Small drawbacks: clicking the menu also activates the active
state of the title. The only workaround I currently see would be to move
the menubar out of the title. And: active state does not look perfect on
search outline. But I think having a feedback is useful for the user.
- Outline-Title-Action is now activated on mouse down instead of click
in order to make it behave the same as the nodes