Table: don't propagate start cell edit key stroke
When pressing ctrl-enter the cell editor opens. This key stroke should
not be propagated. Use case: there may be a key stroke which closes the
form when pressing ctrl-enter. This would prevent the start cell edit
key stroke from working because the form closes before the cell editor
can open.

215283