Table: don't throw exception when endCellEdit cannot be called
CompleteCellEdit is called when the editor should be closed (e.g. using
Enter). After completing the cell edit, endCellEdit will be called which
closes the popup. If there is no popup anymore because completeCellEdit
already closed it (maybe even closed the whole form), endCellEdit must
not fail because there is nothing to do anymore.

Concrete case: CompleteCellEdit closed and reopened the form and failed
because no adapter was found for the given fieldId (form and its
children were already disposed).